+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New Features:
 
 * Thargoid War system progress is now displayed as a progress bar on the in-game overlay when in a TW active system.
+* An activity indicator now briefly flashes green on the overlay when BGS-Tally logs BGS or TW activity.
 
 
 ## v3.1.0-a2 - 2023-08-05